Feat team leader page (#29)
* add team lead menu added in admin

* tema leader list and create team leader page added

* team leader page added

* assign HR to team leader feature added

* column added for role to change HR and TL

* present hr leads of TL, api not working

* withdraw leads api not working and status update of hr api not working

* welcome content updated for team leader and HE

* small warning removed

* conflicts removed

* removed error

* add team leader page tested

* team leader page checked and its inner page , only change TL api remaining

* HR List checked

* rest menu checked , of present HR and not approved HR

* modal corrected and values of tl and admin

* withdraw added in present HR page

* closed leads emptied while marking attendance

* added role as hr when approving employee

* message from backend is showing

* changing team leder and assigning team leader working

* team leader hr table order corrected

* active member, and all mmeber table arranged

* icons improved of sidebar, of evry employee

* table name modified

* aws api added

* error message corrected

* only hr can mark attendance

* handled the case when user in localstorage is undefined

* present count of hr is shown for team leader, and employee attendance show

* attendance button state managed , and appearance improved

* attendace button working

* message added of backend

---------

+++ src/features/common/components/ConfirmationModalBody.js
@@ -9,10 +9,10 @@ function ConfirmationModalBody({ extraObject, closeModal }) {
   //   const [loading, setLoading] = useState(false);
   const dispatch = useDispatch();
 
-  const { message, type, index, params } = extraObject;
+  const { message, type, index, contact, params } = extraObject;
 
   const proceedWithYes = async () => {
-    if (type === CONFIRMATION_MODAL_CLOSE_TYPES.MERGE_WEBSITE_LEADS) {
+    if (type === CONFIRMATION_MODAL_CLOSE_TYPES.WITHDRAW_LEADS) {
       try {
         const storedToken = localStorage.getItem("accessToken");
 
@@ -24,12 +24,92 @@ function ConfirmationModalBody({ extraObject, closeModal }) {
               Authorization: `Bearer ${accessToken}`,
             };
 
-            await axios.post(`${API}/lead/webLeadToLeads`,{}, {
-              headers
+            const response = await axios.post(
+              `${API}/lead/takeLeadsBack`,
+              {
+                contact: contact,
+              },
+              {
+                headers,
+              }
+            );
+
+            console.log("response", response.data);
+
+            dispatch(sliceLeadDeleted(true));
+            dispatch(
+              showNotification({
+                message: `${response.data.message}`,
+                status: 1,
+              })
+            );
+          }
+        } else {
+          dispatch(
+            showNotification({ message: "Access token not found", status: 0 })
+          );
+        }
+      } catch (error) {
+        console.log("error", error);
+        dispatch(
+          showNotification({
+            message: `${error.response.data.message}`,
+            status: 0,
+          })
+        );
+      }
+    } else if (type === CONFIRMATION_MODAL_CLOSE_TYPES.MERGE_WEBSITE_LEADS) {
+      try {
+        const storedToken = localStorage.getItem("accessToken");
+
+        if (storedToken) {
+          const accessToken = JSON.parse(storedToken).token;
+
+          if (accessToken) {
+            const headers = {
+              Authorization: `Bearer ${accessToken}`,
+            };
+
+           const response =  await axios.post(
+              `${API}/lead/webLeadToLeads`,
+              {},
+              {
+                headers,
+              }
+            );
+
+            dispatch(sliceLeadDeleted(true));
+            dispatch(showNotification({ message: `${response.data.message}`, status: 1 }));
+          }
+        } else {
+          dispatch(
+            showNotification({ message: `Access token not found`, status: 0 })
+          );
+        }
+      } catch (error) {
+        dispatch(
+          showNotification({ message: `${error.response.data.message}`, status: 0 })
+        );
+      }
+    } else if (type === CONFIRMATION_MODAL_CLOSE_TYPES.DELETE_ALL_LEAD) {
+      try {
+        const storedToken = localStorage.getItem("accessToken");
+
+        if (storedToken) {
+          const accessToken = JSON.parse(storedToken).token;
+
+          if (accessToken) {
+            const headers = {
+              Authorization: `Bearer ${accessToken}`,
+            };
+
+           const response = await axios.delete(`${API}/lead/delete/bulk`, {
+              headers,
+              params: params,
             });
 
             dispatch(sliceLeadDeleted(true));
-            dispatch(showNotification({ message: "Lead Merged!", status: 1 }));
+            dispatch(showNotification({ message: `${response.data.message}`, status: 1 }));
           }
         } else {
           dispatch(
@@ -38,12 +118,10 @@ function ConfirmationModalBody({ extraObject, closeModal }) {
         }
       } catch (error) {
         dispatch(
-          showNotification({ message: "Error Merging Lead", status: 0 })
+          showNotification({ message: `${error.response.data.message}`, status: 0 })
         );
       }
-    }
-
-    if (type === CONFIRMATION_MODAL_CLOSE_TYPES.DELETE_ALL_LEAD) {
+    } else if (type === CONFIRMATION_MODAL_CLOSE_TYPES.LEAD_DELETE) {
       try {
         const storedToken = localStorage.getItem("accessToken");
 
@@ -55,27 +133,23 @@ function ConfirmationModalBody({ extraObject, closeModal }) {
               Authorization: `Bearer ${accessToken}`,
             };
 
-            await axios.delete(`${API}/lead/delete/bulk`, {
+            const response = await axios.delete(`${API}/lead/${index}`, {
               headers,
-              params: params,
             });
-
             dispatch(sliceLeadDeleted(true));
-            dispatch(showNotification({ message: "Lead Deleted!", status: 1 }));
+            dispatch(showNotification({ message: `${response.data.message}`, status: 1 }));
           }
         } else {
           dispatch(
-            showNotification({ message: "Access token not found", status: 1 })
+            showNotification({ message: "Access token not found", status: 0 })
           );
         }
       } catch (error) {
         dispatch(
-          showNotification({ message: "Error deleting Lead", status: 1 })
+          showNotification({ message: `${error.response.data.message}`, status: 0 })
         );
       }
-    }
-
-    if (type === CONFIRMATION_MODAL_CLOSE_TYPES.LEAD_DELETE) {
+    } else if (type === CONFIRMATION_MODAL_CLOSE_TYPES.MEMBER_DELETE) {
       try {
         const storedToken = localStorage.getItem("accessToken");
 
@@ -87,51 +161,22 @@ function ConfirmationModalBody({ extraObject, closeModal }) {
               Authorization: `Bearer ${accessToken}`,
             };
 
-            await axios.delete(`${API}/lead/${index}`, {
-              headers,
-            });
-            dispatch(sliceLeadDeleted(true));
-            dispatch(showNotification({ message: "Lead Deleted!", status: 1 }));
-          }
-        } else {
-          dispatch(
-            showNotification({ message: "Access token not found", status: 1 })
-          );
-        }
-      } catch (error) {
-        dispatch(
-          showNotification({ message: "Error deleting Lead", status: 1 })
-        );
-      }
-    }
-    if (type === CONFIRMATION_MODAL_CLOSE_TYPES.MEMBER_DELETE) {
-      try {
-        const storedToken = localStorage.getItem("accessToken");
-
-        if (storedToken) {
-          const accessToken = JSON.parse(storedToken).token;
-
-          if (accessToken) {
-            const headers = {
-              Authorization: `Bearer ${accessToken}`,
-            };
-
-            await axios.delete(`${API}/employee/${index}`, {
+            const response = await axios.delete(`${API}/employee/${index}`, {
               headers,
             });
             dispatch(sliceMemberDeleted(true));
             dispatch(
-              showNotification({ message: "Employee Deleted!", status: 1 })
+              showNotification({ message: `${response.data.message}`, status: 1 })
             );
           }
         } else {
           dispatch(
-            showNotification({ message: "Access token not found", status: 1 })
+            showNotification({ message: "Access token not found", status: 0 })
           );
         }
       } catch (error) {
         dispatch(
-          showNotification({ message: "Error deleting employee", status: 1 })
+          showNotification({ message: `${error.response.data.message}`, status: 0 })
         );
       }
     }
